refactor: groups 경로 수정
