+++ src/middleware.ts
@@ -21,7 +21,7 @@ export default function middleware(request: NextRequest) {
       return NextResponse.redirect(new URL('/auth/signin', request.url));
     }
   } else if (pathname.startsWith('/auth')) {
-    return NextResponse.redirect(new URL('/group/groups', request.url));
+    return NextResponse.redirect(new URL('/groups', request.url));
   }
 
   // 기본적으로 요청을 그대로 반환
